# contracts/openapi/mercatus-v1.yaml
openapi: 3.1.0
info:
  title: AIBank — riel de pago para agentes
  version: "1.0.0"
  summary: Autorizar y capturar pagos entre sub-cuentas de una plataforma de agentes.
  description: |
    Implementación del contrato `aibank_integration_contract.md` v0.3 de Mercatus,
    con los defaults acordados en `docs/10-preguntas-mercatus.md` y confirmados en
    la respuesta de Mercatus.

    ## Unidad monetaria

    **Todo importe es un entero en millonésimas (10⁻⁶) de la unidad mayor.**

    | Importe | Valor |
    |---|---|
    | $0.001 | `1000` |
    | $0.50 | `500000` |
    | $1.00 | `1000000` |

    Seis decimales es el máximo representable. Un importe con más precisión se
    rechaza en vez de truncarse en silencio.

    ## Dos modelos de integración

    | | **A — ómnibus** | **B — cuenta propia** |
    |---|---|---|
    | Titular de la cuenta pagadora | La plataforma | El cliente final del banco |
    | Quién autoriza | La plataforma, con su credencial | El titular, con un mandato |
    | Cómo se indica | Sin `mandate_id` | Con `mandate_id` |

    **El contrato de pago es idéntico en los dos.** Lo único que cambia es de
    dónde sale la autoridad para mover el dinero, y se expresa con un campo. Ver
    `POST /v1/consent-requests` para el flujo de permiso del modelo B.

    ## Autorizar y después capturar

    `POST /v1/authorizations` reserva el dinero en la cuenta pagadora sin
    entregarlo. `POST /v1/authorizations/{id}/capture` lo mueve de verdad. Entre
...

...

...

...

...

...
servers:
...
security:
  - bearerToken: []
...
paths:
...
  /v1/consent-requests/{consentRequestId}:
    get:
      tags: [Consentimiento]
      summary: Consulta si el titular ya decidió
      operationId: getConsentRequest
      description: |
        Se consulta, no se notifica: un webhook obligaría a la plataforma a
        exponer un endpoint y al banco a firmar y reintentar, para una decisión
        que tarda lo que tarda una persona en mirar su teléfono.

        Cuando `status` es `approved`, la respuesta trae el `mandate_id` con el
        que se puede empezar a pagar.
      security:
        - bearerToken: [payments:read]
      parameters:
        - name: consentRequestId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Estado de la solicitud.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsentRequest" }
              example:
                consent_request_id: 9c2f1e88-4a7b-4c2d-8e1f-0a1b2c3d4e5f
                handoff_code: csr_kQ8vN2xR7mP4wL9tY6bH3jF5sD1gA0zC
                status: approved
                expires_at: "2026-09-01T20:57:00Z"
                mandate_id: 4a7b1c2d-9e8f-4a1b-8c2d-3e4f5a6b7c8d
                account_id: acc_3f9e8d7c6b5a4938271605f4e3d2c1b0
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: |
            `consent_request_not_found`. Misma respuesta si no existe o si
            pertenece a otra integración.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Motivo legible por máquina. Estable; compáralo por igualdad.
        message:
          type: string
          description: Explicación para un humano. NO es estable; no lo parsees.
...
    ConsentRequest:
      type: object
      required: [consent_request_id, handoff_code, status, expires_at]
      properties:
        consent_request_id: { type: string }
        handoff_code:
          type: string
          description: |
            Lo que el titular abre en su app. Entrégaselo en un enlace o un QR.
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        expires_at: { type: string, format: date-time }
        mandate_id:
          type: string
          description: Presente solo cuando `status` es `approved`.
        account_id:
          type: string
          description: |
            La cuenta que el titular eligió. Presente solo cuando se aprobó.
...
  responses:
    Unauthorized:
      description: |
        `invalid_credential`: falta el token, expiró o no es válido. Pide uno nuevo.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        `insufficient_scope`: el token es válido pero no incluye el scope
        necesario. Pedir otro token con el mismo `client_id` NO lo resuelve.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: |
        `rate_limited`. Espera lo que indique `Retry-After` y reintenta con la
        MISMA `Idempotency-Key`.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...